moved definition functions
